+++ phpstan.neon.dist
@@ -37,7 +37,7 @@ parameters:
         # for src/Schema/TestCase.php
         - '~^Access to an undefined property Atk4\\Data\\Persistence::\$connection\.$~'
         - '~^Call to an undefined method Atk4\\Data\\Persistence::dsql\(\)\.$~'
-        # for src/FieldSqlExpression.php
+        # for src/Field/SqlExpression.php
         - '~^Call to an undefined method Atk4\\Data\\Model::expr\(\)\.$~'
         # for src/Model.php
         - '~^Call to an undefined method Atk4\\Data\\Persistence::update\(\)\.$~'
