+++ src/Schema/Migration.php
@@ -6,7 +6,6 @@ namespace Atk4\Data\Schema;
 
 use Atk4\Data\Exception;
 use Atk4\Data\Field;
-use Atk4\Data\FieldSqlExpression;
 use Atk4\Data\Model;
 use Atk4\Data\Persistence;
 use Atk4\Data\Persistence\Sql\Connection;
@@ -180,7 +179,7 @@ class Migration
         $this->table($model->table);
 
         foreach ($model->getFields() as $field) {
-            if ($field->never_persist || $field instanceof FieldSqlExpression) {
+            if ($field->never_persist || $field instanceof Field\SqlExpression) {
                 continue;
             }
